:recycle: refactor(api): move translateBillingMode() to CertificationCandidate
